Add files via upload
The sales supervisor of a company wants to control the inventory and retail sales of a certain perfume that he distributes in different points of sale, known as "kiosks". For each one of them has the initial inventory and the final inventory of the perfume in question, considering as sales the difference between both inventories.

Develop a visual object-oriented application that allows you to capture this data and calculate the sales of the perfume in each of the kiosks, as well as the amount of the sale based on a price per ounce of the perfume; In addition, you must total both inventories and sales.